Don't bother unlinking /var/lib/netcfg/dhcp6c-finished unless we're using dhcp6c
Rebased-and-modified-by: Philipp Kern <pkern@debian.org>

+++ autoconfig.c
@@ -385,7 +385,8 @@ stop:
 		}
 	}
 
-	unlink(DHCP6C_FINISHED);
+	if (dhcpv6_client == DHCP6C)
+		unlink(DHCP6C_FINISHED);
 
 	return ret;
 }
